CDthread source code revision 2016
This is the first commit of CDthread source code. This code as started as writing in the far 2012. In the while many bug was eliminated and some features was added. At today I have used it in many project but there is a great problem with documentation that is not present ad a separated document. In the fact the only documentation is present in the initial part of the header, and some source, file. Please se the cdthread.h to understand how the code work